backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
langchain-core==0.1.12
langchain-groq==0.0.2
numpy==1.24.3
chromadb==0.4.15
pydantic==2.4.2
python-multipart==0.0.6
sentence-transformers==2.2.2
langchain-text-splitters==0.0.1
